requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
Jinja2==3.1.6
python-multipart==0.0.20
PyMuPDF==1.26.4
Pillow==11.3.0
RapidFuzz==3.13.0
jieba==0.42.1
pytest==8.4.1
httpx==0.28.1
